data-scraper/requirements.txt
openai>=1.0.0
pinecone-client>=2.2.0
python-dotenv>=1.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0 